refactor: OnboardingPage에서만 관련 로직 처리를 할 수 있도록 변경
