+++ app/src/main/java/com/teamwss/websoso/ui/onboarding/model/OnboardingPage.kt
@@ -1,10 +1,28 @@
 package com.teamwss.websoso.ui.onboarding.model
 
-data class OnboardingPage(
+class OnboardingPage(
     val progressPercent: Int,
     val isBackButtonVisible: Boolean,
     val isSkipTextVisible: Boolean
 ) {
+    fun nextPage(): OnboardingPage {
+        val currentIndex = pages.indexOf(this)
+        return if (currentIndex in pages.indices - 1) {
+            pages[currentIndex + 1]
+        } else {
+            this
+        }
+    }
+
+    fun previousPage(): OnboardingPage {
+        val currentIndex = pages.indexOf(this)
+        return if (currentIndex > 0) {
+            pages[currentIndex - 1]
+        } else {
+            this
+        }
+    }
+
     companion object {
         val FIRST = OnboardingPage(
             progressPercent = 33,
